add new ath fields
