+++ packages/examples/package.json
@@ -7,7 +7,9 @@
 		"build": "tsc",
 		"clean": "rimraf dist",
 		"start": "tsx src/index.ts",
-		"start:asset": "tsx src/asset.ts",
+		"start:asset": "tsx src/asset/asset.ts",
+		"start:asset:ath": "tsx src/asset/ath.ts",
+		"start:asset:roi": "tsx src/asset/roi.ts",
 		"start:ai": "tsx src/ai.ts",
 		"start:diligence": "tsx src/diligence.ts",
 		"start:fundraising": "tsx src/fundraising.ts",
